requirements.txt
customtkinter>=5.2.2
yt-dlp>=2025.1.1
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
sounddevice>=0.5.0
soundfile>=0.12.0
